learn-openai/pyproject.toml
[project]
name = "learn-openai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
]
